Adding mention of authorship
Plus a note on the 'hackaround' in the previous commit, e1f1a1263177d22f30932f7932c1b14120bd12e6 : it worked: moving the files to another folder results in deletion of the extant files and upload of the desired files. Then the files were moved back into the right folder. And all is for the best!