Terminology: *glottal cycles* rather than *periods*
There is no check on periodicity (or quasi-periodicity). "Glottal cycles" is therefore a more suitable label than "periods" (a rough first-pass translation from 2004 that had not been corrected since).

Also, added some display of information on the Matlab terminal when the script is run, to make the user aware of some settings for which version 2 of peakdet/peakdet_inter sets default values (inside the script) whereas version 1 asked for user input every time.