Bug fix for signal reinterpolation
The vector of time points needs to have <resampC> (for: resampling Coefficient) times more points than the original (for instance 100 times). The values cover the same time range as the original signal. Thus, the time vector starts at <resampC>, not at 1: 

timevector = [resampC:resampC * LP]

so that when the <timevector> is divided in the next line, the first value will be back to one (1), as desired for reinterpolation.

Prior to this bug fix, there was a short portion of *extrapolated* signal at the very beginning. In theory, this could cause trouble in case this extrapolated portion had an extreme slope.